move class into the cc
